requirements.txt
Django==1.8.1
django-jsonfield==0.9.13
djangorestframework==3.1.1
Markdown==2.6.2
[32m----Using exectuable .venv/bin/pip[0m
